+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## v51.x / v52 (In progress)
+- PC sprites: Added Blanka, Chun-Li, Dhalsim, E. Honda, Guile, Ken, Ryu, Zangief, Balrog, M. Bison, Sagat, Vega, Col. Saturn, Larice, Zarbol, Niiue, Sabitsuki, and Kanji Tatsumi.
+
 ## v51 (2022-05-31)
 - Credits: EBSS AC 2022 winner temple added.
 - Holiday: St. Patrick's Day and Easter options added.